% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{aggregate_composition_tails}
\alias{aggregate_composition_tails}
\title{Aggregate the tails of a matrix}
\usage{
aggregate_composition_tails(M, lb = 1, ub)
}
\arguments{
\item{M}{A \code{matrix}.}

\item{lb}{The row of the matrix to aggregate into.}

\item{ub}{The row of the matrix to aggregate into.}
}
\value{
A \code{matrix}.
}
\description{
Aggregate the tails of a matrix
}
