% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{cut_midpoint}
\alias{cut_midpoint}
\title{Find the midpoints of a vector of bins}
\usage{
cut_midpoint(x)
}
\arguments{
\item{x}{The \code{brms} model fit.}
}
\value{
a \code{function} containing initial values.
}
\description{
Find the midpoints of a vector of bins
}
