% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{data_to_wide}
\alias{data_to_wide}
\title{Data to wide.}
\usage{
data_to_wide(
  x,
  names_from = "length",
  values_from = "count",
  sum_to_one = FALSE,
  ...
)
}
\arguments{
\item{x}{The \code{data.frame} to .}

\item{names_from}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}

\item{values_from}{Other arguments passed on to \code{coord_sf}.}

\item{...}{Other arguments passed on to \code{coord_sf}.}
}
\value{
a \code{data.frame} with the data appended as a matrix.
}
\description{
Data to wide.
}
