% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\docType{data}
\name{default_priors}
\alias{default_priors}
\title{Default priors.}
\format{
An object of class \code{brmsprior} (inherits from \code{data.frame}) with 2 rows and 10 columns.
}
\usage{
default_priors
}
\value{
a \code{brmsprior} \code{data.frame}.
}
\description{
Default priors.
}
\keyword{datasets}
