% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_init}
\alias{get_init}
\title{Get initial values.}
\usage{
get_init(fit)
}
\arguments{
\item{fit}{The \code{brms} model fit.}
}
\value{
a \code{function} containing initial values.
}
\description{
Get initial values.
}
