% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_init_data}
\alias{get_init_data}
\title{Get initial values from the data.}
\usage{
get_init_data(x)
}
\arguments{
\item{fit}{The \code{brms} model fit.}
}
\value{
a \code{function} containing initial values.
}
\description{
Get initial values from the data.
}
