% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{logistic}
\alias{logistic}
\title{Logistic function}
\usage{
logistic(q, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{The \code{brms} model fit.}
}
\value{
a \code{function} containing initial values.
}
\description{
Logistic function
}
