% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(p, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{The \code{brms} model fit.}
}
\value{
a \code{function} containing initial values.
}
\description{
Logit function
}
