% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fixef}
\alias{plot_fixef}
\title{Plot fixed effects.}
\usage{
plot_fixef(
  fits,
  labels = "Males",
  effect = "Intercept",
  xlab = "Midpoint of tail width (mm)"
)
}
\arguments{
\item{fits}{A \code{list} of \code{brmsfit} objects.}

\item{labels}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}

\item{effect}{Other arguments passed on to \code{coord_sf}.}
}
\value{
a \code{ggplot}.
}
\description{
Plot fixed effects.
}
