% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-posterior-predict.R
\name{plot_posterior_predict}
\alias{plot_posterior_predict}
\title{Plot fixed effects.}
\usage{
plot_posterior_predict(object, ndraws = 1000, fit_colour = "blue")
}
\arguments{
\item{fits}{A \code{list} of \code{brmsfit} objects.}

\item{labels}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}

\item{effect}{Other arguments passed on to \code{coord_sf}.}
}
\value{
a \code{ggplot}.
}
\description{
Plot fixed effects.
}
