% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-posterior-predict.R
\name{plot_threshold}
\alias{plot_threshold}
\title{Plot thresholds.}
\usage{
plot_threshold(object, years = c(1990, 1999, 2006, 2012, 2018, 2019), ncol = 2)
}
\arguments{
\item{fits}{A \code{list} of \code{brmsfit} objects.}

\item{labels}{The coordinate reference system to use: integer with the EPSG code, or character with \code{proj4string}.}

\item{effect}{Other arguments passed on to \code{coord_sf}.}
}
\value{
a \code{ggplot}.
}
\description{
Plot thresholds.
}
