% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{aggregate_composition_tails}
\alias{aggregate_composition_tails}
\title{Aggregate the tails of a matrix}
\usage{
aggregate_composition_tails(M, lb = 1, ub)
}
\arguments{
\item{fit}{The \code{brms} model fit.}
}
\value{
a \code{function} containing initial values.
}
\description{
Aggregate the tails of a matrix
}
