% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{data_to_wide}
\alias{data_to_wide}
\title{Data to wide.}
\usage{
data_to_wide(
  x,
  names_from = "length",
  values_from = "count",
  levels = NULL,
  sum_to_one = FALSE
)
}
\arguments{
\item{x}{A \code{data.frame}.}

\item{names_from}{The column name to change to wide format.}

\item{values_from}{The values to change to wide format.}

\item{levels}{The factor levels if wanting to pad missing values with zeros.}

\item{sum_to_one}{Force the data in each row to sum to one.}
}
\value{
a \code{data.frame} with the data in wide format appended as a matrix.
}
\description{
Data to wide.
}
